Inject credentials in environment for online materialize
